HSEARCH-2867 Move JMXRegistar to an exported package in -engine
So that OSGi won't complain that -orm is using a class from a
non-exported package.
